Planet and enemy saving
Added the ability to save enemy and planet locations and then read them out of the save file on game load. Enemies will remove themselves from the save list when they die and so wont be present on next game load assume that the file is saved after the enemy is killed.