Correcting some kind of issue with Fungus
There are two meta files that seem to be missing, but which are not reflected in the github repository.

Assets/Fungus/Integrations/Playmaker/Fungus-PlayMaker.unitypackage.meta

and

Assets/Fungus/Integrations/Spine/Fungus-Spine.unitypackage.meta

I'm not sure what happened here, but I am going to try and update to fix it.