Ability modifications
Updated the laser function to reflect laser charging. Added UI elements to reflect laser charaging and modified the laser controller so that the laser has dimished range following a shot. Range gradually increases as laser recharges.

Also added the foundations of a jump ability to move the ship rapidly from one point to another.

Needed to modify the click manager with a wait function and updated the movement controller to put player movement into it's own function.