Added UI component of installing abilities bug fixes
Added the UI components of installing upgrades. Purchasing the rockets of lasers will now cause them to appear in the lower hud. Also fixed a pair of bugs with the laser ability. When the laser recharge is greater than 1 it was possible to get into a situation where the loop that set the UI elemnts for the laser charge would go out of bounds. This has been fixed. There was also an issue where the laser highlighting wouldn't turn off properly at the end of combat. I believe this has also been resolved.