Added functionality for planets to track resource collection
Added a boolean to the planet object class which determines if the planet has been mined or not. Each planet can now read this boolean and set a parameter on itself to determine if resources have been collected from them or not. If resources have already been collected, then the play gets a warning and nothing more happens.

Also added a new object to act as the star gate and created a script to control its behaviour. These are both blank.