Created prefab for star gate message and connected buttons
Created and formatted a prefab for the star gate message then created the UI function to dsiplay this message. Needed to do several unique things to get this to work including setting the parent of the message, reseting it's scale and transform position, and delegating functions to on click events in the buttons.