Creating star gate prefab and setting up the systems for creating and saving it
Created a prefab for the star gate object and modified the map manager script so it could be spawned and placed correctly each time the map loads. Also modified the UI controller so that it only displays the new game text on truly new games and created the foundation to track which level of solar system the player has reached.