Player death bug
Fixed a bug that allowed the player to reload a map after they have died due to a save operation being triggered by the game quit function. The same operation has been moved to a different location in the script so that it only triggers on a voluntary exit from the game.