Planet threat adjustment, transition timing, and planet spawn bug fixes
Increased the amount of threat that planets add to threat level. Changed the transition timing for scene transitions. Fixed two bugs related to removing entries from lists while iterating through those lists. First, when a planet and warp gate both spawn within one unit of the player's spawn point, I tried to remove the plaent's spawn point from the list of available spawn point for the warp gate. This caused an issue becuase I was removing the item while iterating through that list that held that item. This changed the length of that list and broke the process that was iterating through it. Second, when the player reaches level ten, platinum planets start to spawn, but I want only one of these per solar system, so I removed them from the list of available planets after it spawned. However, I removed it from the wrong list, which caused an index mismatch and the operation failed to complete.