Updated pause menu, modified new enemy, added UI, and changed planet threat values
Changed the pause menu to be click based interactions instead of keyboard interactions. Added an explicit pause button icon to the game screen. Added a quit and save button to the pause menu. Changed the enemy jump behaviour to be partially random. Doubled the amount that landing on planets adds to the threat level.