Updating tutorial level with pause system and new dialog
Adding all the elements of the new pause system to the tutorial level and making sure they are hooked up correctly. Also added an effect that ensures the pause menu is always on top of all other UI elements when it is open. Lastly, changed the tutorial text to explain maxing out the threat level system.