Correcting visual glitches on main menu
The High Score screen had a bug that would put the button on top of the scores. This was fixed by forcing the components into more rigidly defined coordinates. The Credits screen was difficult to read because the text was too small and the scroll bar was difficult to interact with.