Updating tutorial level, fixing enemy range display, removing new game message
Updating the tutorial level so that any references to the mouse or clicking are replaced with touch screen appropriate statements.

Fixed the enemy range display so that it now works properly in all contexts. Tapping on an enemy will show their attack range, tapping again will turn it off. Tapping anywhere else will also turn off the display. Highlighting is also properly handled when turning on weapons or abilities with range displays.

Finally, removed the new game message from the main game as this will no longer be a test build.