Updating resource depleted message, dialog menu button ordering, and creating a windows icon
Changed the fade rate on the resources depleted message so it stays on the screen longer. Also changed the text from "All resources have been collected from this planet" to "resources depleted".

Updated the UI in the tutorial level so that dialog buttons are no longer on top of the pause menu.

Created an icon image to use for the windows build of the game.